physfsrwops: fix compilation error by fixing the functions' signatures

+++ util/physfsrwops.c
@@ -27,7 +27,7 @@ static const char *get_error(void)
     return PHYSFS_getErrorByCode(PHYSFS_getLastErrorCode());
 }
 
-static int physfsrwops_seek(SDL_RWops *rw, int offset, int whence)
+static long physfsrwops_seek(SDL_RWops *rw, long offset, int whence)
 {
     PHYSFS_file *handle = (PHYSFS_file *) rw->hidden.unknown.data1;
     int pos = 0;
@@ -100,7 +100,7 @@ static int physfsrwops_seek(SDL_RWops *rw, int offset, int whence)
 } /* physfsrwops_seek */
 
 
-static int physfsrwops_read(SDL_RWops *rw, void *ptr, int size, int maxnum)
+static size_t physfsrwops_read(SDL_RWops *rw, void *ptr, size_t size, size_t maxnum)
 {
     PHYSFS_file *handle = (PHYSFS_file *) rw->hidden.unknown.data1;
     PHYSFS_sint64 rc = PHYSFS_readBytes(handle, ptr, size * maxnum);
@@ -114,7 +114,7 @@ static int physfsrwops_read(SDL_RWops *rw, void *ptr, int size, int maxnum)
 } /* physfsrwops_read */
 
 
-static int physfsrwops_write(SDL_RWops *rw, const void *ptr, int size, int num)
+static size_t physfsrwops_write(SDL_RWops *rw, const void *ptr, size_t size, size_t num)
 {
     PHYSFS_file *handle = (PHYSFS_file *) rw->hidden.unknown.data1;
     PHYSFS_sint64 rc = PHYSFS_writeBytes(handle, ptr, size * num);
